Implement system controller

+++ src/main/java/com/creants/creants_2x/socket/gate/IQAntUser.java
@@ -35,7 +35,7 @@ public interface IQAntUser {
 	String getScreenSize();
 
 
-	long getUserId();
+	int getUserId();
 
 
 	String getUserName();
